Reorganisation des onglets clients
