+++ src/app/pages/clients/entreprise/liste-entreprise/liste-entreprise.component.html
@@ -1,0 +1,1 @@
+<p>liste-entreprise works!</p>
