+++ src/app/pages/clients/entreprise/liste-entreprise/liste-entreprise.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ListeEntrepriseComponent } from './liste-entreprise.component';
+
+describe('ListeEntrepriseComponent', () => {
+  let component: ListeEntrepriseComponent;
+  let fixture: ComponentFixture<ListeEntrepriseComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ ListeEntrepriseComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(ListeEntrepriseComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
