+++ src/app/pages/clients/entreprise/liste-entreprise/liste-entreprise.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-liste-entreprise',
+  templateUrl: './liste-entreprise.component.html',
+  styleUrls: ['./liste-entreprise.component.scss']
+})
+export class ListeEntrepriseComponent {
+
+}
